+++ src/main/java/com/jeontongju/seller/mapper/SellerMapper.java
@@ -1,6 +1,8 @@
 package com.jeontongju.seller.mapper;
 
 import com.jeontongju.seller.domain.Seller;
+import com.jeontongju.seller.dto.temp.SellerInfoDto;
+import com.jeontongju.seller.dto.temp.SellerInfoForAuctionDto;
 import com.jeontongju.seller.dto.temp.SignUpInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -20,5 +22,23 @@ public class SellerMapper {
         .businessmanName(signUpInfo.getBusinessmanName())
         .businessmanPhoneNumber(signUpInfo.getBusinessmanPhoneNumber())
         .build();
+  }
+
+  public SellerInfoDto toSellerInfoDto(Seller seller) {
+    return SellerInfoDto.builder()
+            .storeName(seller.getStoreName())
+            .storeImageUrl(seller.getStoreImageUrl())
+            .build();
+  }
+
+  public SellerInfoForAuctionDto toSellerInfoForAuctionDto(Seller seller) {
+
+    return SellerInfoForAuctionDto.builder()
+            .storeImageUrl(seller.getStoreImageUrl())
+            .storeName(seller.getStoreName())
+            .storeEmail(seller.getEmail())
+            .storePhoneNumber(seller.getStorePhoneNumber())
+            .businessmanName(seller.getBusinessmanName())
+            .build();
   }
 }
