+++ src/main/java/com/jeontongju/seller/service/SellerService.java
@@ -4,9 +4,9 @@ import com.jeontongju.seller.domain.Seller;
 import com.jeontongju.seller.dto.reqeust.ModifySellerInfo;
 import com.jeontongju.seller.dto.reqeust.SellerJudgeRequestDto;
 import com.jeontongju.seller.dto.response.*;
-
 import com.jeontongju.seller.dto.response.SellerInfoForConsumerDto;
 import com.jeontongju.seller.dto.temp.SellerInfoDto;
+import com.jeontongju.seller.dto.temp.SellerInfoForAuctionDto;
 import com.jeontongju.seller.dto.temp.SignUpInfo;
 import com.jeontongju.seller.exception.SellerEntityNotFoundException;
 import com.jeontongju.seller.kafka.SellerProducer;
@@ -33,7 +33,7 @@ public class SellerService {
 
   public SellerInfoDto getSellerInfo(Long sellerId) {
 
-    return SellerInfoDto.toDto(
+    return sellerMapper.toSellerInfoDto(
         sellerRepository.findById(sellerId).orElseThrow(SellerEntityNotFoundException::new));
   }
 
@@ -91,7 +91,7 @@ public class SellerService {
         sellerRepository.findById(memberId).orElseThrow(SellerEntityNotFoundException::new);
     seller.modifySeller(modifySellerInfo);
 
-    sellerProducer.sendUpdateSeller(SellerInfoDto.toDto(seller));
+    sellerProducer.sendUpdateSeller(sellerMapper.toSellerInfoDto(seller));
   }
 
   @Transactional
@@ -101,10 +101,10 @@ public class SellerService {
 
   public SellerInfoForAuctionDto getSellerInfoForAuction(Long sellerId) {
 
-    return SellerInfoForAuctionDto.toDto(
+    return sellerMapper.toSellerInfoForAuctionDto(
         sellerRepository.findById(sellerId).orElseThrow(SellerEntityNotFoundException::new));
   }
- 
+
   public List<GetSellerByAdminDto> getSellerListByAdmin() {
 
     return sellerRepository.findAll().stream()
